Menghapus dan melakukan refactoring Large class
Menghapus bad small dan melakukan refactoring pada class Employee menjadi  4 class yaitu class GajiPegawai, DataPribadi, KeluargaPegawai, TahunPegawai
