+++ TaxCalculator/src/lib/DataPribadi.java
@@ -1,0 +1,10 @@
+package lib;
+
+public class DataPribadi extends Employee{
+    private String employeeId;
+	private String firstName;
+	private String lastName;
+	private String idNumber;
+	private String address; 
+
+}
