+++ TaxCalculator/src/lib/KeluargaPegawai.java
@@ -1,0 +1,18 @@
+package lib;
+
+public class KeluargaPegawai extends Employee {
+    private String spouseName;
+	private String spouseIdNumber;
+	private List<String> childNames;
+	private List<String> childIdNumbers;
+
+    public void setSpouse(String spouseName, String spouseIdNumber) {
+		this.spouseName = spouseName;
+		this.spouseIdNumber = idNumber;
+	}
+	
+	public void addChild(String childName, String childIdNumber) {
+		childNames.add(childName);
+		childIdNumbers.add(childIdNumber);
+	}
+}
