+++ TaxCalculator/src/lib/TahunPegawai.java
@@ -1,0 +1,11 @@
+package lib;
+
+public class TahunPegawai extends Employee{
+
+    private int yearJoined;
+	private int monthJoined;
+	private int dayJoined;
+	private int monthWorkingInYear;
+
+    
+}
